# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bunkarr API",
    "version": "1",
    "description": "Bunkarr REST API. Authenticate with the API key from Settings > General in the X-Api-Key header (or the apikey query parameter), or with a login session cookie. Errors are {\"message\": \"...\"}: 400 invalid input, 401 not authenticated, 403 refused (a cross-site browser request, an action that needs a login session, a directory Bunkarr cannot read), 404 unknown id, 409 conflict with the current state (including safety refusals that need a confirmation or a mounted share), 502 Plex failed, 503 unavailable (the backup routes of a server without backup services, or a request stopped as the server shuts down), 500 anything else. Secrets (Plex tokens, Apprise URLs) are write-only. Job-starting endpoints answer 202 with the queued Job. Sizes are bytes, times RFC 3339.",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "apiKeyHeader": []
    },
    {
      "apiKeyQuery": []
    },
    {
      "session": []
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "apiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bunkarr_session"
      }
    },
    "schemas": {
...
      "Destination": {
        "type": "object",
        "required": [
          "id",
          "name",
          "engine",
          "target",
          "enabled",
          "sourceIds",
          "fsType",
          "capabilities",
          "settings",
          "retention",
          "schedule",
          "verifySchedule",
          "lastJob",
          "lastSync",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "engine": {
            "type": "string",
            "enum": [
              "filecopy"
            ]
          },
          "target": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "sourceIds": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "fsType": {
            "type": "string"
          },
          "capabilities": {
            "$ref": "#/components/schemas/Capabilities"
          },
          "settings": {
            "$ref": "#/components/schemas/DestinationSettings"
          },
          "retention": {
            "$ref": "#/components/schemas/Retention"
          },
          "schedule": {
            "$ref": "#/components/schemas/CronSchedule"
          },
          "verifySchedule": {
            "$ref": "#/components/schemas/CronSchedule"
          },
          "lastJob": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Job"
              },
              {
                "type": "null"
              }
            ],
            "description": "The newest job of any type that worked on the destination (sync, verify, retention, Plex DB backup)"
          },
          "lastSync": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Job"
              },
              {
                "type": "null"
              }
            ],
            "description": "The newest sync of the destination (queued, running or finished; previews excluded): its last sync status. null before the first sync"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DestinationInput": {
        "type": "object",
        "description": "Create: name and target are required. Update: empty or missing fields keep the stored values.",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "engine": {
            "type": "string",
            "enum": [
              "filecopy"
            ]
          },
          "target": {
            "type": "string",
            "description": "Absolute path of the mounted share; it must exist. Cannot change after creation."
          },
          "enabled": {
            "type": "boolean"
          },
          "sourceIds": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "settings": {
            "$ref": "#/components/schemas/DestinationSettings"
          },
          "retention": {
            "$ref": "#/components/schemas/Retention"
          },
          "schedule": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CronSchedule"
              }
            ],
            "description": "Sync schedule. Create: missing means none. Update: missing keeps it. {cron: \"\", enabled: false} removes it."
          },
          "verifySchedule": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CronSchedule"
              }
            ],
            "description": "Verify schedule. Create: missing means 0 5 * * 0 (weekly, Sunday 05:00). Update: missing keeps it."
          },
          "attach": {
            "type": "boolean",
            "description": "Create only: adopt the existing .bunkarr/destination.json marker (S3)."
          },
          "allowLocal": {
            "type": "boolean",
            "description": "Create only: accept a target on a local filesystem (S3)."
          }
        }
      },
      "DestinationTest": {
        "type": "object",
        "required": [
          "target"
        ],
        "properties": {
          "target": {
            "type": "string"
          }
        }
      },
      "DestinationTestResult": {
        "type": "object",
        "required": [
          "ok",
          "marker",
          "writable",
          "fsType",
          "local",
          "capabilities",
          "freeBytes",
          "totalBytes",
          "entries",
          "message",
          "warnings"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "marker": {
            "type": "string",
            "enum": [
              "ok",
              "missing",
              "mismatch",
              "foreign"
            ]
          },
          "writable": {
            "type": "boolean"
          },
          "fsType": {
            "type": "string"
          },
          "local": {
            "type": "boolean"
          },
          "capabilities": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Capabilities"
              },
              {
                "type": "null"
              }
            ]
          },
          "freeBytes": {
            "type": "integer"
          },
          "totalBytes": {
            "type": "integer"
          },
          "entries": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "SyncRequest": {
        "type": "object",
        "properties": {
          "dryRun": {
            "type": "boolean",
            "description": "Scan, plan and guard only; nothing is written (S9)"
          },
          "allowChanges": {
            "type": "boolean",
            "description": "Run changes the mass-change guard holds (S10b)"
          }
        }
      },
      "VerifyRequest": {
        "type": "object",
        "properties": {
          "dryRun": {
            "type": "boolean"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "No such object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflicts with the current state (the message says why)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal error (message redacted of secrets)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "No backup services are configured on this server, or the request was stopped (the server is shutting down)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
...
    }
  },
  "paths": {
...
    "/destinations": {
      "get": {
        "operationId": "listDestinations",
        "summary": "List destinations",
        "responses": {
          "200": {
            "description": "Destinations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Destination"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Destinations"
        ]
      },
      "post": {
        "operationId": "createDestination",
        "summary": "Add a destination: checks and probes the target and writes its marker (S3); creates its schedules",
        "description": "A target that already holds a marker needs attach (409 otherwise); a target on a local filesystem needs allowLocal (400 otherwise).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DestinationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destination"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Destinations"
        ]
      }
    },
    "/destinations/test": {
      "post": {
        "operationId": "testDestinationTarget",
        "summary": "Check a target before creating a destination (writes nothing)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DestinationTest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationTestResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Destinations"
        ]
      }
    },
    "/destinations/{id}": {
      "get": {
        "operationId": "getDestination",
        "summary": "Get a destination",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Destination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destination"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Destinations"
        ]
      },
      "put": {
        "operationId": "updateDestination",
        "summary": "Update a destination and its schedules (the target cannot change)",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DestinationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destination"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Destinations"
        ]
      },
      "delete": {
        "operationId": "deleteDestination",
        "summary": "Forget a destination and its schedules; the backup data at the target is never touched",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Done"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Destinations"
        ]
      }
    },
    "/destinations/{id}/test": {
      "post": {
        "operationId": "testDestination",
        "summary": "Check a destination: marker, filesystem, free space; re-runs the capability probe",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationTestResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Destinations"
        ]
      }
    },
    "/destinations/{id}/sync": {
      "post": {
        "operationId": "syncDestination",
        "summary": "Sync the destination's sources now (dryRun: preview only)",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The queued job (an identical queued job is returned instead of a new one)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Destinations"
        ]
      }
    },
    "/destinations/{id}/verify": {
      "post": {
        "operationId": "verifyDestination",
        "summary": "Verify the destination's files now",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The queued job (an identical queued job is returned instead of a new one)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Destinations"
        ]
      }
    },
    "/destinations/{id}/snapshots": {
      "get": {
        "operationId": "listSnapshots",
        "summary": "Plex database versions stored at the destination, newest first",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshots",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Snapshot"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "tags": [
          "Destinations"
        ]
      }
    },
...
  }
}